Fixes  #5: Set constraint on filter button so it can not be invoked before solution info is set, delete FilterState enum since we do not need to keep track of solution info set on filter, a
