+++ GitBranchDiffer/GitBranchDiffer.csproj
@@ -63,7 +63,6 @@
     <Compile Include="GitBranchDifferPluginOptions.cs">
       <SubType>Component</SubType>
     </Compile>
-    <Compile Include="Filter\FilterInitializationState.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="GitBranchDifferPackage.cs" />
     <Compile Include="GitBranchDifferValidator.cs" />
