+++ GitBranchDiffer/GitBranchDifferPackage.vsct
@@ -56,6 +56,8 @@
       <Button guid="guidBranchDiffPackageCmdSet" id="cmdidBranchDiffAndFilter" priority="0x0400" type="Button">
         <Parent guid="guidSHLMainMenu" id="IDG_VS_TOOLBAR_PROJWIN_FILTERS" />
         <Icon guid="guidDiffFilterImage" id="GitBranchDifferDiffIcon16" />
+        <CommandFlag>DefaultInvisible</CommandFlag>
+        <CommandFlag>DynamicVisibility</CommandFlag>
         <Strings>
           <ButtonText>Show diff in solution</ButtonText>
           <CommandName>cmdidBrandDiffAndFilter</CommandName>
@@ -73,6 +75,10 @@
       <Bitmap guid="guidDiffFilterImage" href="Resources\GitBranchDifferIconList_16_20.png" usedList="GitBranchDifferDiffIcon16" />
     </Bitmaps>
   </Commands>
+
+  <VisibilityConstraints>
+      <VisibilityItem guid="guidBranchDiffPackageCmdSet" id="cmdidBranchDiffAndFilter" context="UICONTEXT_SolutionExists"/>
+  </VisibilityConstraints>
 
   <Symbols>
     <!-- This is the package guid. -->
